refactor(utils): move userWithId function to discord subdirectory to improve code organization and readability update: update import statements in affected files to reflect new location of userWithId function
refactor: move userWithId function to discord utils folder and update imports in affected files to use the new path
