+++ src/commands/forum/[sub-commands]/resolve.cmd.ts
@@ -5,7 +5,7 @@ import { msgParams } from "$core/utils/message";
 import { commands } from "$core/configs/message/command";
 import { ChannelType, ForumChannel } from "discord.js";
 import { logger } from "$core/utils/logger";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 import { trustDiscordLinks } from "../forum.const";
 
 export const execute: CommandExecute = async(command) => {
