+++ src/commands/main/[sub-commands]/group-remove/remove-role.cmd.ts
@@ -5,7 +5,7 @@ import { msgParams } from "$core/utils/message";
 import { gqlRequest } from "$core/utils/request";
 import { commands } from "$core/configs/message/command";
 import { logger } from "$core/utils/logger";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 
 export const execute: CommandExecute = async(command) => {
   const role = command.options.getRole(commands.main.groups.remove.subcmds.role.options.role.name, true);
