+++ src/commands/stats/stats.cmd.ts
@@ -10,7 +10,7 @@ import { gqlRequest } from "$core/utils/request";
 import { ChartJSNodeCanvas } from "chartjs-node-canvas";
 import { AttachmentBuilder, EmbedBuilder } from "discord.js";
 import { logger } from "$core/utils/logger";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 
 export const execute: CommandExecute = async(command) => {
   // Get server activity :
