+++ src/commands/top-voice/[sub-commands]/month.cmd.ts
@@ -8,7 +8,7 @@ import { msgParams } from "$core/utils/message";
 import { gqlRequest } from "$core/utils/request";
 import { memberPerPage } from "../top-voice.const";
 import { logger } from "$core/utils/logger";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 
 export const execute: CommandExecute = async(command) => {
   let members: MembersData[] = [];
