+++ src/events/guild-member-add/guild-member-add.event.ts
@@ -8,7 +8,7 @@ import { logger } from "$core/utils/logger";
 import { msgParams } from "$core/utils/message";
 import { gqlRequest } from "$core/utils/request";
 import { ChannelType } from "discord.js";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 
 export const event: EventName = "guildMemberAdd";
