+++ src/events/user-update/user-update.event.ts
@@ -2,7 +2,7 @@ import type { EventExecute, EventName } from "$core/utils/handler/event";
 import { gqlRequest } from "$core/utils/request";
 import { setUsernameAndProfilePicture } from "$core/api/requests/member";
 import { logger } from "$core/utils/logger";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 
 export const event: EventName = "userUpdate";
