+++ src/tasks/verify-members/verify-members.task.ts
@@ -5,7 +5,7 @@ import { getGuildMembers } from "$core/utils/discord/guild";
 import type { TaskExecute, TaskInterval } from "$core/utils/handler/task";
 import { logger } from "$core/utils/logger";
 import { gqlRequest } from "$core/utils/request";
-import { userWithId } from "$core/utils/user";
+import { userWithId } from "$core/utils/discord/user";
 
 export const interval: TaskInterval = "0 */3 * * * *";
